+++ package.json
@@ -22,8 +22,20 @@
     "@fortawesome/react-fontawesome": "^0.1.13",
     "@material-ui/core": "^4.11.2",
     "@material-ui/icons": "^4.11.2",
-    "@types/express": "^4.17.11",
-    "@types/node": "^12.0.0",
+    "express": "^4.17.1",
+    "node-fetch": "^2.6.1",
+    "react": "^17.0.1",
+    "react-cookie-consent": "^6.2.1",
+    "react-dom": "^17.0.1",
+    "react-ga": "^3.3.0",
+    "react-hook-form": "^6.14.0",
+    "react-material-ui-carousel": "^2.1.2",
+    "react-refresh": "^0.9.0",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "4.0.1",
+    "react-transition-group": "^4.4.1"
+  },
+  "devDependencies": {
     "@types/react": "^16.9.53",
     "@types/react-dom": "^16.9.8",
     "@types/react-router-dom": "^5.1.6",
@@ -37,23 +49,10 @@
     "eslint-plugin-jsx-a11y": "latest",
     "eslint-plugin-react": "^7.21.5",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "express": "^4.17.1",
     "gh-pages": "^3.1.0",
     "husky": "^4.3.6",
     "lint-staged": "^10.5.3",
-    "node-fetch": "^2.6.1",
     "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-cookie-consent": "^6.2.1",
-    "react-dom": "^17.0.1",
-    "react-ga": "^3.3.0",
-    "react-hook-form": "^6.14.0",
-    "react-material-ui-carousel": "^2.1.2",
-    "react-refresh": "^0.9.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
-    "react-transition-group": "^4.4.1",
-    "serve": "^11.3.2",
     "stylelint": "^13.8.0",
     "stylelint-config-standard": "^20.0.0",
     "typescript": "^4.0.3"
