+++ src/server/server.js
@@ -1,12 +1,13 @@
 const express = require('express');
 const path = require('path');
-const {fetchMembers} = require("./updateMembers");
+const { fetchMembers, submitContactForm } = require("./requests");
 
 const apiKey = process.env.AIRTABLE_API_KEY || process.argv[2];
 const PORT = process.env.PORT || 3042;
 const app = express();
+app.use(express.text());
 
-// cache members array to minimize hubspot api calls; update every 10 minutes
+// cache members array; update every 5 minutes
 let membersCache;
 fetchMembers(apiKey)
   .then(members => membersCache = members)
@@ -15,11 +16,16 @@ setInterval(() => {
   fetchMembers(apiKey)
     .then(members => membersCache = members)
     .catch(error => console.log(error))
-}, 600000);
+}, 300000);
 
 // path to fetch dOrg active builder data
 app.get("/members", (request, response) => {
   response.json(membersCache);
+});
+
+// path to submit contact form
+app.post("/submitContactForm", (request, response) => {
+  response.json(submitContactForm(apiKey, request.body));
 });
 
 // send react app for all other get requests
